Context-sensitive res holograms; Improved logging; Some formatting improvements
Res holograms now display differently depending on whether a player owns a residence at that point or not.

The custom logging layout now properly displays exception data. Previously, only the exception class of the top-level exception cause was displayed. Now, this is shown for each stack trace available.

A central instance of EntityHider is now stored in the main Dragons instance (previously this was a static method in the NPC class)

Some additional testing commands were added to assist in debugging.

Continuing to make some minor improvements to code formatting.