Guild Update
Extended base functionality for guilds, which are named groups of players which compete for experience points, resources, etc. Guilds can be public, invite-only, request-only, or unlisted.

Players can belong to multiple guilds. Under the current architecture, guilds are created by administrators but run by players. This is to ensure that guilds are compatible with the game plotline and to regulate quantity and quality. The guild owner, who is not necessarily a server admin, has complete authority to moderate the guild and manage membership as they see fit.

Fixed a bug in GuildChannelHandler that incorrectly bound it to the Staff channel rather than the Guild channel. Whoops!

Also added a /mongo command for administrators to view basic database statistics regarding connectivity and collection enumeration.

Minor changes to dragons-core to expose certain fields of the MongoDB connection string and to use a less technical error message in StringUtil#parseEnum.