Guild bugfixes and more devtools
Fixed bugs with the /guild command related to argument parsing.

Fixed a bug that queried the database unnecessarily when reliance on local pooling was appropriate.

Added a /panic command to crash the server when in an error state. A full dump will be emitted to the log and the server will be shut down as gracefully as possible: first through the default Bukkit mechanism, and if that fails, by forcibly exiting the runtime. When the server needs to be shut down immediately without saving, for example if regularly persisted data has become corrupted without a backup, the /crashserver command is still necessary to avoid any auto-save or cleanup.